% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/choropleth.R
\name{map_choropleth}
\alias{map_choropleth}
\title{Update a map with a choropleth overlay}
\usage{
map_choropleth(session, map_ID, tileset, fill_colour, pickable = TRUE)
}
\arguments{
\item{session}{<\code{shiny::session}> The Shiny session object.}

\item{map_ID}{<\code{character}> A unique identifier for the map input.}

\item{tileset}{<\code{character}> The tileset to be used for the choropleth overlay.}

\item{fill_colour}{<\code{data.frame}> A tibble with two columns: 'ID' and 'fill'. ID is
the ID of the feature, and fill are hexes of 6 digits.}

\item{pickable}{<\code{logical}> Should there be hovered effect, indicating the layer
can be pickable? Defaults to TRUE.}
}
\value{
No return value. The function sends an update message to the Shiny
server to update the map.
}
\description{
This function updates a map with a choropleth overlay. It generates a
configuration list that includes the tileset and fill colour for the choropleth
overlay and sends this configuration to the server to update the map.
}
