% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/choropleth.R
\name{map_choropleth_update_fill_colour}
\alias{map_choropleth_update_fill_colour}
\title{Update a map's choropleth overlay fill colour}
\usage{
map_choropleth_update_fill_colour(session, map_ID, fill_colour)
}
\arguments{
\item{session}{<\code{shiny::session}> The Shiny session object.}

\item{map_ID}{<\code{character}> A unique identifier for the map input.}

\item{fill_colour}{<\code{data.frame}> A tibble with two columns: 'ID' and 'fill'. ID is
the ID of the feature, and fill are hexes of 6 digits.}
}
\value{
No return value. The function sends an update message to the Shiny
server to update the map.
}
\description{
This function updates a map's choropleth overlay fill colour. It generates a
configuration list that includes the fill colour for the choropleth overlay
and sends this configuration to the server to update the map.
}
