% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map.R
\name{map_input}
\alias{map_input}
\title{Create a React Shiny input for a map}
\usage{
map_input(
  map_ID,
  username,
  token,
  longitude,
  latitude,
  zoom,
  tileset_prefix,
  div_height = "100vh",
  div_width = "100\%"
)
}
\arguments{
\item{map_ID}{<\code{character}> A unique identifier for the map input.}

\item{username}{<\code{character}> Mapbox username, where the tilesets live.}

\item{token}{<\code{character}> Necessary token to access mapbox.}

\item{longitude}{<\code{numeric}> The longitude value for the initial map center.}

\item{latitude}{<\code{numeric}> The latitude value for the initial map center.}

\item{zoom}{<\code{numeric}> The zoom level for the initial map display.}

\item{tileset_prefix}{<\code{numeric}> The prefix for the tileset to be used with the map.
This will be used only for the stories tileset.}

\item{div_height}{<\code{character}> Height of the input. The map will take 100\%
of the space. Defaults to the entire viewport (100vh).}

\item{div_width}{<\code{character}> Width of the input. The map will take 100\%
of the space. Defaults to 100\% of the width of the viewport (100\%).}
}
\value{
A React Shiny input widget for the map.
}
\description{
This function creates a React Shiny input for a map. It generates a map
input widget with specified parameters and adds it to the Shiny application.
}
